requirements.txt
aioredis==1.3.1
asgiref==3.2.5
astroid==2.2.5
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
autopep8==1.4.4
boto3==1.12.23
botocore==1.15.23
certifi==2019.6.16
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.3
django-allauth==0.39.1
django-channels==0.7.0
django-cors-headers==3.0.2
django-countries==5.5
django-heroku==0.3.1
django-rest-auth==0.9.5
django-storages==1.9.1
djangorestframework==3.9.4
docutils==0.15.2
gunicorn==19.9.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
jmespath==0.9.5
lazy-object-proxy==1.4.1
mccabe==0.6.1
msgpack==0.6.2
oauthlib==3.0.1
pep8==1.7.1
Pillow==6.2.0
psycopg2==2.7.2
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.3.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.3.3
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.0
stripe==2.43.0
Twisted==20.3.0
txaio==20.1.1
typed-ast==1.4.0
urllib3==1.25.3
whitenoise==5.0.1
wrapt==1.11.2
zope.interface==5.0.1
